Add support for variable RngRun values
This commit modifies the get_space and get_results_as_* functions so that it's
possible to extract results even if the number of runs to extract is not
specified. The output space is filled with NaN values, in order to have all
'runs' axes of the same size (corresponding to the maximum number of randomized
repetitions that were performed among all parameter combinations).